+++ Sources/EZNetworking/Interceptors/Protocols/DownloadTaskInterceptor.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+/// Protocol for intercepting download tasks specifically
+public protocol DownloadTaskInterceptor: AnyObject {
+    func urlSession(_ session: URLSession, downloadTask: URLSessionDownloadTask, didFinishDownloadingTo location: URL)
+    func urlSession(_ session: URLSession, downloadTask: URLSessionDownloadTask, didWriteData bytesWritten: Int64, totalBytesWritten: Int64, totalBytesExpectedToWrite: Int64)
+    func urlSession(_ session: URLSession, downloadTask: URLSessionDownloadTask, didResumeAtOffset fileOffset: Int64, expectedTotalBytes: Int64)
+}
