--- Sources/EZNetworking/Interceptors/Protocols/Interface.swift
@@ -1,69 +1,0 @@
-import Foundation
-
-/// Protocol for intercepting URL cache operations
-public protocol CacheInterceptor: AnyObject {
-    func urlSession(_ session: URLSession, dataTask: URLSessionDataTask, willCacheResponse proposedResponse: CachedURLResponse, completionHandler: @escaping @Sendable (CachedURLResponse?) -> Void)
-    
-    func urlSession(_ session: URLSession, dataTask: URLSessionDataTask, willCacheResponse proposedResponse: CachedURLResponse) async -> CachedURLResponse?
-}
-
-/// Protocol for intercepting URL authentication challenges
-public protocol AuthenticationInterceptor: AnyObject {
-    func urlSession(_ session: URLSession, task: URLSessionTask, didReceive challenge: URLAuthenticationChallenge, completionHandler: @escaping @Sendable (URLSession.AuthChallengeDisposition, URLCredential?) -> Void)
-    
-    func urlSession(_ session: URLSession, didReceive challenge: URLAuthenticationChallenge, completionHandler: @escaping @Sendable (URLSession.AuthChallengeDisposition, URLCredential?) -> Void)
-    
-    func urlSession(_ session: URLSession, task: URLSessionTask, didReceive challenge: URLAuthenticationChallenge) async -> (URLSession.AuthChallengeDisposition, URLCredential?)
-    
-    func urlSession(_ session: URLSession, didReceive challenge: URLAuthenticationChallenge) async -> (URLSession.AuthChallengeDisposition, URLCredential?)
-}
-
-/// Protocol for intercepting URL redirect operations
-public protocol RedirectInterceptor: AnyObject {
-    func urlSession(_ session: URLSession, task: URLSessionTask, willPerformHTTPRedirection response: HTTPURLResponse, newRequest request: URLRequest, completionHandler: @escaping @Sendable (URLRequest?) -> Void)
-    
-    func urlSession(_ session: URLSession, task: URLSessionTask, willPerformHTTPRedirection response: HTTPURLResponse, newRequest request: URLRequest) async -> URLRequest?
-}
-
-/// Protocol for intercepting URL metrics collection
-public protocol MetricsInterceptor: AnyObject {
-    func urlSession(_ session: URLSession, task: URLSessionTask, didFinishCollecting metrics: URLSessionTaskMetrics)
-}
-
-/// Protocol for intercepting task lifecycle events
-public protocol TaskLifecycleInterceptor: AnyObject {
-    func urlSession(_ session: URLSession, task: URLSessionTask, didCompleteWithError error: Error?)
-    func urlSession(_ session: URLSession, taskIsWaitingForConnectivity task: URLSessionTask)
-    func urlSession(_ session: URLSession, didCreateTask task: URLSessionTask)
-}
-
-/// Protocol for intercepting data tasks specifically
-public protocol DataTaskInterceptor: AnyObject {
-    func urlSession(_ session: URLSession, dataTask: URLSessionDataTask, didReceive response: URLResponse, completionHandler: @escaping @Sendable (URLSession.ResponseDisposition) -> Void)
-    func urlSession(_ session: URLSession, dataTask: URLSessionDataTask, didReceive data: Data)
-    func urlSession(_ session: URLSession, dataTask: URLSessionDataTask, didBecome downloadTask: URLSessionDownloadTask)
-    func urlSession(_ session: URLSession, dataTask: URLSessionDataTask, didBecome streamTask: URLSessionStreamTask)
-    
-    func urlSession(_ session: URLSession, dataTask: URLSessionDataTask, didReceive response: URLResponse) async -> URLSession.ResponseDisposition
-}
-
-/// Protocol for intercepting download tasks specifically
-public protocol DownloadTaskInterceptor: AnyObject {
-    func urlSession(_ session: URLSession, downloadTask: URLSessionDownloadTask, didFinishDownloadingTo location: URL)
-    func urlSession(_ session: URLSession, downloadTask: URLSessionDownloadTask, didWriteData bytesWritten: Int64, totalBytesWritten: Int64, totalBytesExpectedToWrite: Int64)
-    func urlSession(_ session: URLSession, downloadTask: URLSessionDownloadTask, didResumeAtOffset fileOffset: Int64, expectedTotalBytes: Int64)
-}
-
-/// Protocol for intercepting stream tasks specifically
-public protocol StreamTaskInterceptor: AnyObject {
-    func urlSession(_ session: URLSession, readClosedFor streamTask: URLSessionStreamTask)
-    func urlSession(_ session: URLSession, writeClosedFor streamTask: URLSessionStreamTask)
-    func urlSession(_ session: URLSession, betterRouteDiscoveredFor streamTask: URLSessionStreamTask)
-    func urlSession(_ session: URLSession, streamTask: URLSessionStreamTask, didBecome inputStream: InputStream, outputStream: OutputStream)
-}
-
-/// Protocol for intercepting WebSocket tasks specifically
-public protocol WebSocketTaskInterceptor: AnyObject {
-    func urlSession(_ session: URLSession, webSocketTask: URLSessionWebSocketTask, didOpenWithProtocol protocol: String?)
-    func urlSession(_ session: URLSession, webSocketTask: URLSessionWebSocketTask, didCloseWith closeCode: URLSessionWebSocketTask.CloseCode, reason: Data?)
-}
