+++ Sources/EZNetworking/Interceptors/Protocols/RedirectInterceptor.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+/// Protocol for intercepting URL redirect operations
+public protocol RedirectInterceptor: AnyObject {
+    func urlSession(_ session: URLSession, task: URLSessionTask, willPerformHTTPRedirection response: HTTPURLResponse, newRequest request: URLRequest, completionHandler: @escaping @Sendable (URLRequest?) -> Void)
+    
+    func urlSession(_ session: URLSession, task: URLSessionTask, willPerformHTTPRedirection response: HTTPURLResponse, newRequest request: URLRequest) async -> URLRequest?
+}
