+++ Sources/EZNetworking/Interceptors/Protocols/WebSocketTaskInterceptor.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+/// Protocol for intercepting WebSocket tasks specifically
+public protocol WebSocketTaskInterceptor: AnyObject {
+    func urlSession(_ session: URLSession, webSocketTask: URLSessionWebSocketTask, didOpenWithProtocol protocol: String?)
+    func urlSession(_ session: URLSession, webSocketTask: URLSessionWebSocketTask, didCloseWith closeCode: URLSessionWebSocketTask.CloseCode, reason: Data?)
+}
